refactor: x 버튼 눌렀을 때 아이템 전체 삭제되도록 수정
