+++ app/src/main/java/woowacourse/shopping/application/ShoppingApplication.kt
@@ -5,9 +5,9 @@ import woowacourse.shopping.data.ShoppingDatabase
 import woowacourse.shopping.data.cart.repository.CartRepositoryImpl
 import woowacourse.shopping.data.history.repository.HistoryRepositoryImpl
 import woowacourse.shopping.feature.cart.CartViewModel
-import woowacourse.shopping.feature.cart.ViewModelFactory
 import woowacourse.shopping.feature.goods.GoodsViewModel
 import woowacourse.shopping.feature.goodsdetails.GoodsDetailsViewModel
+import woowacourse.shopping.util.ViewModelFactory
 
 class ShoppingApplication : Application() {
     private val database by lazy { ShoppingDatabase.getDatabase(applicationContext) }
